refactor(physics): decouple per-scheme Data and Parameters from echam aggregator
Phase 7. Each typed Data sub-struct now lives next to the scheme that
owns/writes it; the monolithic ECHAM ``Parameters`` aggregator and
``echam_physics_data.py`` are gone. Composition is the only place where
scheme-native Parameters are pulled together — ``echam_physics()`` takes
per-scheme keyword arguments directly.

Data struct moves:
- ``RadiationData``  → ``radiation/radiation_types.py``
                       (next to ``RadiationParameters``).
- ``CloudData``      → ``clouds/cloud_data.py`` (new).
- ``VerticalDiffusionData`` → ``vertical_diffusion/tte_tke/vertical_diffusion_types.py``.
- ``SurfaceData``    → ``surface/echam/surface_types.py``
                       (already held ``SurfaceParameters``).
- ``AerosolData``    → ``aerosol/aerosol_types.py`` (new).
- ``ChemistryData``  → ``chemistry/simple_chemistry.py``
                       (next to ``ChemistryParameters``).
- ``DiagnosticData`` and ``PhysicsData`` were dead — deleted with the
  file.

Factory and Hydra plumbing:
- ``echam_physics()`` no longer accepts a monolithic ``Parameters``
  argument. It takes per-scheme kwargs (``convection``, ``clouds``,
  ``microphysics``, ``microphysics_2m``, ``radiation``,
  ``vertical_diffusion``, ``surface``, ``aerosol``, ``hines``,
  ``sso``); each ``None`` resolves to the scheme's ``.default()``.
- ``runners._build_echam_param_kwargs`` resolves
  ``cfg.physics.params.<subgroup>`` directly into those kwargs without
  any aggregator round-trip. The CLI surface
  (``physics.params.convection.entrpen=4e-4``) is unchanged.
- ``parameters_test.py`` rewritten against the per-scheme API.
- ``notebooks/06_macv2_aerosols.py`` and the lazy ``Parameters`` import
  in ``echam/__init__.py`` updated.

Files deleted:
- ``jcm/physics/echam/parameters.py``
- ``jcm/physics/echam/echam_physics_data.py``

All 822 fast tests pass. Lint passes.
