+++ jcm/physics/chemistry/simple_chemistry.py
@@ -64,11 +64,76 @@ class ChemistryState(NamedTuple):
 
 class ChemistryTendencies(NamedTuple):
     """Tendencies from chemistry processes"""
-    
+
     # Trace gas tendencies (mixing ratio per second)
     ozone_tend: jnp.ndarray         # Ozone tendency (ppbv/s)
     methane_tend: jnp.ndarray       # Methane tendency (ppbv/s)
     co2_tend: jnp.ndarray          # CO2 tendency (ppmv/s)
+
+
+@tree_math.struct
+class ChemistryData:
+    """Diagnostic sub-struct for the chemistry diagnostic-dict slot.
+
+    Seeded by ``EchamBoundaryConditions`` (which fills CO2/CH4/O3 VMRs
+    from forcing) and consumed by the radiation terms; the
+    ``SimpleChemistry`` term also writes its production/loss diagnostics
+    here. Lives next to the chemistry scheme so a future replacement
+    chemistry term can extend or replace this struct without reaching
+    into the ECHAM tree.
+    """
+
+    # Gas concentrations (volume mixing ratios)
+    ozone_vmr: jnp.ndarray           # Ozone VMR [ppbv] (nlev, ncols)
+    methane_vmr: jnp.ndarray         # Methane VMR [ppbv] (nlev, ncols)
+    co2_vmr: jnp.ndarray            # CO2 VMR [ppmv] (nlev, ncols)
+
+    # Production/loss rates
+    ozone_production: jnp.ndarray    # Ozone production rate [ppbv/s] (nlev, ncols)
+    ozone_loss: jnp.ndarray         # Ozone loss rate [ppbv/s] (nlev, ncols)
+    methane_loss: jnp.ndarray       # Methane loss rate [ppbv/s] (nlev, ncols)
+
+    # Surface emissions/deposition
+    methane_surface_flux: jnp.ndarray  # Surface methane flux [ppbv m/s] (ncols,)
+    ozone_dry_deposition: jnp.ndarray  # Ozone dry deposition velocity [m/s] (ncols,)
+
+    @classmethod
+    def zeros(cls, nodal_shape, nlev):
+        # Chemistry fields should be in column format (nlev, ncols)
+        # where ncols = nlat * nlon.
+        if len(nodal_shape) == 2:
+            nlat, nlon = nodal_shape
+            ncols = nlat * nlon
+            column_shape = (nlev, ncols)
+            surface_shape = (ncols,)
+        else:
+            column_shape = (nlev,) + nodal_shape
+            surface_shape = nodal_shape
+
+        return cls(
+            ozone_vmr=jnp.zeros(column_shape),
+            methane_vmr=jnp.zeros(column_shape),
+            co2_vmr=jnp.zeros(column_shape),
+            ozone_production=jnp.zeros(column_shape),
+            ozone_loss=jnp.zeros(column_shape),
+            methane_loss=jnp.zeros(column_shape),
+            methane_surface_flux=jnp.zeros(surface_shape),
+            ozone_dry_deposition=jnp.zeros(surface_shape),
+        )
+
+    def copy(self, **kwargs):
+        new_data = {
+            'ozone_vmr': self.ozone_vmr,
+            'methane_vmr': self.methane_vmr,
+            'co2_vmr': self.co2_vmr,
+            'ozone_production': self.ozone_production,
+            'ozone_loss': self.ozone_loss,
+            'methane_loss': self.methane_loss,
+            'methane_surface_flux': self.methane_surface_flux,
+            'ozone_dry_deposition': self.ozone_dry_deposition,
+        }
+        new_data.update(kwargs)
+        return ChemistryData(**new_data)
 
 
 def compute_height_from_pressure(
