+++ jcm/physics/radiation/rrtmgp.py
@@ -27,8 +27,8 @@ from jax_solar import OrbitalTime, radiation_flux, get_solar_sin_altitude
 from jcm.physics.radiation.radiation_types import (
     RadiationParameters,
     RadiationTendencies,
+    RadiationData,
 )
-from jcm.physics.echam.echam_physics_data import RadiationData
 from jcm.physics.radiation.grey_two_stream.radiation_scheme import prepare_radiation_state
 from jcm.physics.radiation.cloud_optics import (
     effective_radius_liquid,
